fix: renamed the attention directory to avoid conflict with other formats
