--- src/quantum_attention/__init__.py
@@ -1,4 +1,0 @@
-from .run_attention import run_q_attention
-__all__ = [
-    "run_q_attention"
-]
